# fix: improve grants list performance