## Summary

<!-- Include a summary of the change and/or why it's necessary. -->

Grants list queries the server for each user or group which is exceedingly slow when more than a few users/groups exist. It's also unnecessary in a lot of use cases because users and groups will be duplicated for multiple grants. Instead, retrieve a list of users and groups once and lookup the user when necessary

Old code:

```
$ /usr/bin/time -h infra grants list | grep USER
  USER (7702)        ACCESS         DESTINATION
        20.20s real             1.21s user              0.82s sys
```

New code:

```
$ /usr/bin/time -h infra grants list | grep USER
  USER (7702)        ACCESS         DESTINATION
        0.28s real              0.14s user              0.05s sys
```

Tested against local Infra server with a permutation of ~100 users, 7 roles, and 10 destinations

## Checklist

<!-- 
Checklists help us remember things. Change [ ] to [x] to show completion.
-->

- [ ] Wrote appropriate unit tests
- [ ] Considered security implications of the change
- [ ] Updated associated docs where necessary
- [ ] Updated associated configuration where necessary
- [ ] Change is backwards compatible if it needs to be (user can upgrade without manual steps?)
- [ ] Nothing sensitive logged
- [ ] Considered data migrations for smooth upgrades


## Related Issues

<!--
Link any related issues. Each issue should be on
its own line. For example:

Resolves #1234
Resolves #4321
-->

Resolves #2231